Refactorish the streamish
